+++ src/balance/command.ts
@@ -1,18 +1,15 @@
 import Entropy from "@entropyxyz/sdk"
-import { Base } from "../common/base"
-import { BalanceUtils } from "./utils"
+import { BaseCommand } from "../common/base-command"
+import * as BalanceUtils from "./utils"
 import { FLOW_CONTEXT } from "./constants"
 
-export class BalanceCommand extends Base {
-  private readonly balanceService: BalanceUtils
-
+export class BalanceCommand extends BaseCommand {
   constructor (entropy: Entropy, endpoint: string) {
     super(entropy, endpoint, FLOW_CONTEXT)
-    this.balanceService = new BalanceUtils(this.entropy, endpoint)
   }
 
   public async getBalance (address: string) {
-    const balance = await this.balanceService.getBalance(address)
+    const balance = await BalanceUtils.getBalance(this.entropy, address)
 
     this.logger.log(`Current balance of ${address}: ${balance}`, `${BalanceCommand.name}`)