+++ CHANGELOG.md
@@ -27,8 +27,10 @@ Version header format: `[version] Name - year-month-day (entropy-core compatibil
 - new: 'src/flows/user-program-management/view.ts' - service file for pure functions of viewing user programs
 - new: 'src/flows/user-program-management/helpers/utils.ts' - utility helper file for user program management specific methods
 - new: './src/flows/user-program-management/remove.ts' - service file for removing user program pure function
-- new: './src/common/base.ts' - base abstract class for new command and utils classes
+- new: './src/common/base.ts' - base abstract class for new command classes
 - new: './src/balance' - new file structure for our CLI/TUI flows
+  - new: './src/balance/command.ts' - main entry file for balance command for tui/cli
+  - new: './src/balance/utils.ts' - utilities and helper methods for all things balance
 
 ### Changed
 - folder name for user programs to match the kebab-case style for folder namespace